requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
openpyxl>=3.1.2
python-pptx>=0.6.23
pdfplumber>=0.10.3
tavily-python>=0.3.0
python-dotenv>=1.0.0
